Started to change all of the guitar stuff to bass stuff
Worked on replacing variables, changing function names etc. So that it will play bass sounds rather than the old guitar sounds. The reason I chose to do this is because the old sound pack was inconsistent with each sound. But in this bass sound pack each sound is a specific note.